config: add a name field to config.json
To identify bundles lets add a name field to the config.json with a URL-like
structure. The field should be descriptive and human readable and give a user
indication on the application inside the bundle. We will use a URL-like
structure to facilitate discovery of bundle images in the future. For example
all of the following are all valid names:

    example.com
    example.com/app
    example.com/org/app
    subdomain.example.com/org/project/app

This leaves room for future use cases while giving existing use cases
like Docker images a simple mapping from existing names e.g.
index.docker.io/group/my-app.